alex-gabriel-alves-faustino/projeto-1/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
google-generativeai
google-api-python-client==2.118.0
lyricsgenius